Categories dans GetFilm + add/remove categorie

+++ Controller/Categorie.py
@@ -27,11 +27,14 @@ def getCategorie():
 @route_blueprint.route("/categorie/list", methods=["GET"])
 def getListCategorie():
 	param = request.get_json()
-	page = int(param.get(["page"], 1)) - 1
-	if page is None:
+	try:
+		page = int(param.get("page", 1)) - 1
+		if page is None or page < 0:
+			return jsonify({"status": 422, "message": "Parameters Error"})
+	except ValueError:
 		return jsonify({"status": 422, "message": "Parameters Error"})
 	
-	sql = "SELECT * FROM categorie LIMIT ? OFFSET ?"
+	sql = "SELECT uid, nom FROM categorie LIMIT ? OFFSET ?"
 	data = Database.request(sql, ((page+1)*10, page*10))
 	if data is None:
 		return jsonify({"status": 422, "message": "SQL Error"})
@@ -61,7 +64,14 @@ def deleteCategorie():
 	uid = param["uid"]
 	if uid is None:
 		return jsonify({"status": 422, "message": "Parameters Error"})
-	
+
+	# DELETE film_categorie (enlève la catégorie à tous les films)
+	sql = "DELETE FROM film_categorie WHERE categorie_id = (SELECT id FROM categorie WHERE uid = ?)"
+	data = Database.request(sql, (uid,))
+	if data is None:
+		return jsonify({"status": 422, "message": "SQL Error"})
+
+	# DELETE la categorie
 	sql = "DELETE FROM categorie WHERE uid = ?"
 	data = Database.request(sql, (uid,))
 	if data is None:
